requirements.txt
loguru
websockets-proxy
fake-useragent